improve test runner

+++ test/test.py
@@ -12,9 +12,18 @@ class Common(object):
     TEST_DIR = os.path.join(SRC_DIR, 'test')
     DATA_DIR = os.path.join(SRC_DIR, 'share')
     PDF2HTMLEX_PATH = os.path.join(SRC_DIR, 'pdf2htmlEX')
+
+    SAVE_TMP = os.environ.get('P2H_TEST_SAVE_TMP')
+    CANONICAL_TEMPDIR = '/tmp/pdf2htmlEX_test'
     
     def setUp(self):
-        self.cur_temp_dir = tempfile.mkdtemp(prefix='pdf2htmlEX_test')
+        if not self.SAVE_TMP:
+            self.cur_temp_dir = tempfile.mkdtemp(prefix='pdf2htmlEX_test')
+        else:
+            shutil.rmtree(self.CANONICAL_TEMPDIR, True)
+            os.mkdir(self.CANONICAL_TEMPDIR)
+            self.cur_temp_dir = self.CANONICAL_TEMPDIR
+
         self.cur_data_dir = os.path.join(self.cur_temp_dir, 'share')
         self.cur_output_dir = os.path.join(self.cur_temp_dir, 'out')
         os.mkdir(self.cur_data_dir)
@@ -36,9 +45,12 @@ class Common(object):
         # copy files
         shutil.copy(os.path.join(self.DATA_DIR, 'base.min.css'),
                 os.path.join(self.cur_data_dir, 'base.min.css'))
+        shutil.copy(os.path.join(self.TEST_DIR, 'fancy.min.css'),
+                os.path.join(self.cur_data_dir, 'fancy.min.css'))
 
     def tearDown(self):
-        shutil.rmtree(self.cur_temp_dir, True)
+        if not self.SAVE_TMP:
+            shutil.rmtree(self.cur_temp_dir, True)
 
     def run_pdf2htmlEX(self, args):
         """
@@ -49,19 +61,15 @@ class Common(object):
         :return: an object of relevant info
         """
 
-        cmd = [self.PDF2HTMLEX_PATH,
+        args = [self.PDF2HTMLEX_PATH,
                '--data-dir', self.cur_data_dir,
                '--dest-dir', self.cur_output_dir
-              ]
+              ] + args
 
-        for val in args:
-            cmd.append(str(val))
-
-        return_code = subprocess.call(cmd)
-        self.assertEquals(return_code, 0)
+        return_code = subprocess.call(list(map(str, args)))
+        self.assertEquals(return_code, 0, 'cannot execute pdf2htmlEX')
 
         files = os.listdir(self.cur_output_dir)
-        files.sort()
 
         return { 
             'return_code' : return_code,
@@ -75,9 +83,20 @@ if __name__ == '__main__':
         exit(1)
     suites = []
     loader = unittest.TestLoader()
-    for module_name in ['test_naming']:
+    test_names = list(map(lambda x: 'T.'+x, sys.argv[1:]))
+    for module_name in ['test_naming', 'test_conversion']:
         __import__(module_name)
-        suites.append(loader.loadTestsFromModule(sys.modules[module_name]))
+        if len(test_names) > 0:
+            try:
+                suites.append(loader.loadTestsFromNames(test_names, sys.modules[module_name]))
+            except:
+                pass
+        else:
+            suites.append(loader.loadTestsFromModule(sys.modules[module_name]))
+
+    if len(suites) == 0:
+        print >>sys.stderr, 'No test found'
+        exit(1)
 
     failure_count = 0
     runner = unittest.TextTestRunner(verbosity=2)
